% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{add_inter_clusters_connections}
\alias{add_inter_clusters_connections}
\title{Add connections between cluster nodes}
\usage{
add_inter_clusters_connections(G, col.names, weight.factor)
}
\arguments{
\item{G}{An \code{igraph} object, representing a Scaffold map, as returned by \code{add_vertices_to_landmarks_graph}}

\item{col.names}{A vector of column names to be used for the comptation of similarities}

\item{weight.factor}{Weight factor. The edge weights for the inter-cluster connections will be multiplied
by this weight factor. This is useful in case one wants to weight the connections between the clusters differently
from the connections between the clusters and the landmarks}
}
\value{
Returns an \code{igraph} object, representing a Scaffold map with inter-cluster connections
}
\description{
This function takes an existing Scaffold map and adds connections between the cluster nodes
}
