% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{add_vertices_to_landmarks_graph}
\alias{add_vertices_to_landmarks_graph}
\title{Add cluster vertices to a landmark graph}
\usage{
add_vertices_to_landmarks_graph(
  G,
  tab.clustered,
  tab.landmarks,
  col.names,
  min.similarity = 0.5
)
}
\arguments{
\item{G}{An \code{igraph} object representing the landmark graph}

\item{tab.clustered}{A \code{data.frame} containing the clusters to be represented in the Scaffold map. Each
row corresponds to a cluster, and each column to a different marker}

\item{tab.landmarks}{A \code{data.frame} containing the data corresponding to the landmark nodes. Each row corresponds
to a landmark, and each column to a different cluster}

\item{col.names}{A character vector of column names to be used for the analysis (these must correspond to the names
of the columns in \code{tab.clustered})}

\item{min.similarity}{The minimum similarity value to be used as threshold to filter out edges. See the documentation
for \code{\link{get_distances_from_landmarks}}}
}
\value{
Returns a new \code{igraph} object with the added cluster vertices
}
\description{
Add cluster vertices to a landmark graph
}
