% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{build_graph}
\alias{build_graph}
\title{Builds a graph from a data.frame of nodes}
\usage{
build_graph(tab, col.names, filtering_T = 0.8)
}
\arguments{
\item{tab}{A \code{data.frame} of graph nodes. All the columns of \code{tab} will become vertex properties of the output graph}

\item{col.names}{A character vector indicating which columns of \code{tab} should be used to calculate distances}

\item{filtering_T}{The threshold used to filter edges in the graph. If this value is \code{< 1} the edges are filtered based on the
cosine similarity value. If this is an integer \code{>= 1} they are filtered based on rank for each node (i.e. for each node only
the edges with rank less than the threshold are retained)}
}
\value{
Returns an \code{igraph} object
}
\description{
Builds a graph from a data.frame of nodes
}
