% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceatlas2.R
\name{complete_forceatlas2}
\alias{complete_forceatlas2}
\title{Performs a Complete cycle of ForceAtlas2}
\usage{
complete_forceatlas2(G, first.iter = 1000, overlap.method = NULL,
  overlap.iter = NULL, ...)
}
\arguments{
\item{G}{The input graph}

\item{first.iter}{The maximum number of iterations in the first step, which is performed without overlap resolution}

\item{overlap.method}{If this is \code{NULL} overlap resolution is not performed. Otherwise this should be a string specifying the
overlap resolution method. Two options are possible
\itemize{
  \item{\code{"repel"}}: This is the method used in the original ForceAtlas2 implementation. Using this method, a repulsive force
    is applied to nodes that overlap each other. This method can cause problem in cases where the layout is extremely crowded,
    as this repulsive force becomes the major determinant of the layout, and the nodes end up being arranged essentially in a grid
  \item{\code{"expand"}}: In this method, the graph is linearly expanded, until no two nodes overlap anymore
}}

\item{overlap.iter}{The maximum number of iterations for the overlap resolution step. This is only used if \code{overlap.method} is not
\code{NULL}}
}
\value{
Returns an \code{igraph} object with two additional vertex attributes \code{x} and \code{y}, containing the x and y coordinates
  of the vertices in the final layout
}
\description{
This function performs a complete (i.e. possibly including overlap resolution) cycle of the ForceAtlas2 force-directed layout algorithm
}
