% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{cosine_similarity_from_matrix}
\alias{cosine_similarity_from_matrix}
\title{Calculate cosine similarity between a vector and the rows of a matrix}
\usage{
cosine_similarity_from_matrix(x, m)
}
\arguments{
\item{x}{A numeric vector of length \code{P}}

\item{m}{An \code{NxP} matrix}
}
\value{
Returns a vector of length \code{N} containing the cosine similarity between
  the vector \code{x} and all the rows of \code{m}
}
\description{
Calculate cosine similarity between a vector and the rows of a matrix
}
