% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{cosine_similarity_matrix}
\alias{cosine_similarity_matrix}
\title{Calculate cosine similarity between the rows of a matrix}
\usage{
cosine_similarity_matrix(m)
}
\arguments{
\item{m}{An \code{N x P} matrix}
}
\value{
Returns an \code{N x N} matrix with the cosine similarity between the corresponding rows in \code{m}
}
\description{
This function calculates the cosine similarity between the rows of an input matrix, according to the values
of the variables in the columns
}
