% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{distance_from_attractor_hard_filter}
\alias{distance_from_attractor_hard_filter}
\title{Remove connections to landmark nodes based on an expression threshold}
\usage{
distance_from_attractor_hard_filter(dd, tab, col.names, thresh = 0.5)
}
\arguments{
\item{dd}{A matrix of cosine similarity values between cluster nodes (rows) and landmark nodes (columns)}

\item{tab}{A \code{data.frame} containing the data (median expression values) for the landmark nodes}

\item{col.names}{A vector of colum names to be used for the computation}

\item{thresh}{The threshold to use. Connections to landmarks for which the expression of all the markers
is \code{< thresh}, will be removed}
}
\value{
Returns the same dd matrix, but with the values corresponding to the connections that have been
  removed set to 0
}
\description{
This function removes connections to a landmark node, in cases where the expression of all markers
for that landmark is less than a threshold. This is useful when mapping data that contains markers which
are all expected to be absent in certain landmarks
}
