% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{downsample_by}
\alias{downsample_by}
\title{Downsample an input data.frame}
\usage{
downsample_by(tab, col.name, size)
}
\arguments{
\item{tab}{Input \code{data.frame}}

\item{col.name}{The name of a column in tab (typically the data in the column should either be characters or factors).
For each value of \code{tab$col.name}, there will be at most \code{size} row in the output table}

\item{size}{The number of rows to select for each value of \code{tab$col.name}. If size is larger than the total number of rows
available for a particular value, all rows for that value will be returned}
}
\value{
Returns a \code{data.frame}
}
\description{
This function downsamples an input \code{data.frame}, so that for each value of \code{col.name}, there are
at most \code{size} rows.
}
