% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{filter_matrix}
\alias{filter_matrix}
\title{Filter a matrix based on a threshold value}
\usage{
filter_matrix(m, threshold)
}
\arguments{
\item{treshold}{The threhsold used for filtering. See Description for details}
}
\value{
This function does not return any value. The input matrix is modified in-place
}
\description{
This function filters an input matrix based on a threshold value. Values below
the threshold are set to 0 For each row this function calculates \code{max(row)}.
The actual threshold used for filtering is \code{min(threshold, max(row))}. This guarantees that
there is always at least one non-zero value in each row. WARNING: the input matrix is modified
in-place.
}
