% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{filter_matrix_by_rank}
\alias{filter_matrix_by_rank}
\title{Filter a matrix based on the rank of the values in each row}
\usage{
filter_matrix_by_rank(m, threshold)
}
\arguments{
\item{m}{The input matrix}

\item{threshold}{The threshold rank. Values with rank greater than the threhsold will be set to
0. Note that the rank is 1-based (i.e. the largest observation has rank 1)}
}
\value{
This function does not return any value. The input matrix is modified in-place.
}
\description{
This function filters an input matrix, by ranking the data in each row from largest to smallest,
and setting any element whose rank is greater than the input threshold to 0. In other words,
if the treshold is X, only the X greatest values in each row will be kept, and the rest will
be set to 0. WARNING: the input matrix is modified in-place.
}
