% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{get_distances_from_attractors}
\alias{get_distances_from_attractors}
\title{Calculate (and filter) disatances between cluster and landmark ondes}
\usage{
get_distances_from_attractors(m, tab, col.names, q.thresh, min.thresh)
}
\arguments{
\item{m}{A \code{data.frame} containing expression values for the cluster nodes}

\item{tab}{A \code{data.frame} containing expression values for the landmark nodes}

\item{col.names}{A vector of column names to be used for the computation}

\item{q.thresh}{The quantile probability to be used for filtering edges. The algorithm
will calculate a weight threshold based on this quantile of the weight distribution}

\item{min.thresh}{The minimum value to be used for thresholding edges. Irrespective of the results of the
quantile computation, the actual threshold used will never be less than this value}
}
\value{
Returns a matrix with each row corresponding to a cluster node, and each column
  corresponding to the similarity between the cluster node and the respective landmark node. This matrix
  can be directly used as an adjacency matrix for graph construction
}
\description{
This function calculates the cosine similarity between the cluster and landmark nodes, and then filters
out the resulting matrix to include only the top-scoring edges. The result of this function can be used
as an adjacency matrix for graph construction
}
