% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{get_scaffold_map}
\alias{get_scaffold_map}
\title{Creates a Scaffold map}
\usage{
get_scaffold_map(
  tab.clustered,
  col.names,
  tab.landmarks,
  G.landmarks = NULL,
  ew.influence = ceiling(length(col.names)/3),
  min.similarity = 0.5,
  inter.cluster.col.names = col.names,
  inter.cluster.weight.factor = 0.7,
  overlap.method = "repel"
)
}
\arguments{
\item{tab.clustered}{A \code{data.frame} containing the clusters to be represented in the Scaffold map. Each
row corresponds to a cluster, and each column to a different marker}

\item{col.names}{A character vector of column names to be used for the analysis (these must correspond to the names
of the columns in \code{tab.clustered})}

\item{tab.landmarks}{A \code{data.frame} containing the data corresponding to the landmark nodes. Each row corresponds
to a landmark, and each column to a different cluster}

\item{G.landmarks}{An \code{igraph} object represeting the identity and position of landmark nodes. If this is
\code{NULL} the position of the landmark nodes will be calculated de-novo as part of the computation.}

\item{min.similarity}{The minimum similarity value to be used as threshold to filter out edges. See the documentation
for \code{\link{get_distances_from_landmarks}}}

\item{inter.cluster.col.names}{A vector of column names to be used to calculate inter-cluster edges If this
is \code{NULL} no inter-cluster edges will be present in the Scaffold map (not recommended)}

\item{inter.cluster.weight.factor}{The weight of the inter-cluster edges will be multplied by this number before
calculating the ForceAtlas2 layout. This can be used to fine tune the relative contribution of cluster-to-landmark vs
cluster-to-cluster edges in the laytout}

\item{overlap.method}{The method to be used for resolving nodes overlap in the final layout. See the documentation for
\code{\link{complete_forceatlas2}}}
}
\value{
Returns a list with the following components
 \itemize{
    \item{\code{G.landmarks}}: an \code{igraph} object representing the identity and position of landmark nodes, suitable
      for succesive invocations of this function
    \item{\code{G.complete}}: an \code{igraph} object representing the Scaffold map
  }
}
\description{
Creates a Scaffold map
}
