% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{get_unsupervised_graph}
\alias{get_unsupervised_graph}
\title{Builds an unsupervised force-directed graph}
\usage{
get_unsupervised_graph(
  tab,
  col.names,
  filtering.threshold,
  method = c("forceatlas2", "umap"),
  ...
)
}
\arguments{
\item{tab}{A \code{data.frame} of graph nodes. All the columns of \code{tab} will become vertex properties of the output graph}

\item{col.names}{A character vector indicating which columns of \code{tab} should be used to calculate distances}

\item{filtering.threshold}{The threshold used to filter edges in the graph}

\item{...}{Additional argument passed to \link{build_graph} or \link{build_umap_graph} depending
on the choice of method}
}
\value{
Returns an \code{igraph} graph object. This function also performs graph-based clustering. The resulting cluster membership
  is contained in the \code{community_id} vertex attribute of the resulting graph
}
\description{
Builds an unsupervised force-directed graph
}
