% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{load_attractors}
\alias{load_attractors}
\title{Load landmark populations from a list of FCS files}
\usage{
load_attractors(files.list, asinh.cofactor, transform.data = T, ...)
}
\arguments{
\item{files.list}{The list of files to load. The population names will be derived by splitting the file name
using \code{"_"} as separator, and taking the last field}

\item{asinh.cofactor}{The cofactor to use for \code{asinh} transformation, only used if \code{transform.data} is \code{TRUE}}

\item{transform.data}{Logical, whether \code{asinh} transformation should be applied to the data}

\item{...}{Additional arguments passed to \code{flowCore::read.FCS}}
}
\value{
Returns a list with the following elements:
  \itemize{
    \item{\code{downsampled.data}}: a \code{data.frame} containing the data, downsampled to a fixed number of events
    \item{\code{tab.attractors}}: a \code{data.frame} with the data for the landmark nodes. Each row is a separate landmark
      population, and the columns correspond to the median expression values of each marker
    \item{\code{celltype_key}}: a \code{data.frame} containing the key for the names of the landmark populations. The
      \code{cellType} column is a numeric id corresponding to the column of the same name in \code{tab.attractors}
  }
}
\description{
Load landmark populations from a list of FCS files
}
