% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{load_attractors_from_dir}
\alias{load_attractors_from_dir}
\title{Load landmark populations from a directory}
\usage{
load_attractors_from_dir(dir, ...)
}
\arguments{
\item{dir}{A directory name. All the FCS files in the directory will be loaded}

\item{...}{Arguments passed on to \code{load_attractors}
\describe{
  \item{asinh.cofactor}{The cofactor to use for \code{asinh} transformation, only used if \code{transform.data} is \code{TRUE}}
  \item{transform.data}{Logical, whether \code{asinh} transformation should be applied to the data}
}}
}
\value{
Returns a list with the following elements:
  \itemize{
    \item{\code{downsampled.data}}: a \code{data.frame} containing the data, downsampled to a fixed number of events
    \item{\code{tab.attractors}}: a \code{data.frame} with the data for the landmark nodes. Each row is a separate landmark
      population, and the columns correspond to the median expression values of each marker
    \item{\code{celltype_key}}: a \code{data.frame} containing the key for the names of the landmark populations. The
      \code{cellType} column is a numeric id corresponding to the column of the same name in \code{tab.attractors}
  }
}
\description{
Load landmark populations from a directory
}
