% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{load_landmarks}
\alias{load_landmarks}
\title{Load landmark populations from a list of FCS files}
\usage{
load_landmarks(files.list, asinh.cofactor, transform.data, ...)
}
\arguments{
\item{files.list}{The list of files to load. The population names will be derived by splitting the file name
using \code{"_"} as separator, and taking the last field}

\item{asinh.cofactor}{The cofactor to use for \code{asinh} transformation. If this is \code{NULL} no transformation
is applied}

\item{...}{Additional arguments passed to \code{flowCore::read.FCS}}
}
\value{
Returns a list with the following elements:
  \itemize{
    \item{\code{downsampled.data}}: a \code{data.frame} containing the data, downsampled to a fixed number of events
    \item{\code{tab.landmarks}}: a \code{data.frame} with the data for the landmark nodes. Each row is a separate landmark
      population, and the columns correspond to the median expression values of each marker. The \code{data.frame} also
      contains a column called \code{cellType} which contains the name of the corresponding landmark. In order to derive
      the name of the landmark. the name of the corresponding FCS file is split on the \code{"_"} character, and the last
      field is used for the landmark name.
  }
}
\description{
Load landmark populations from a list of FCS files
}
