% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_graph}
\alias{write_graph}
\title{Write a graph as GraphML file}
\usage{
write_graph(G, out.name, ...)
}
\arguments{
\item{G}{The \code{igraph} object representing the graph}

\item{out.name}{The name of the output file}

\item{...}{Additional arguments passed to \code{igraph::write.graph}}
}
\description{
This function writes a graph as a GraphML file. Please use this function when
writing graphs created using this package
}
