% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{write_landmarks_data}
\alias{write_landmarks_data}
\title{Write landmarks data}
\usage{
write_landmarks_data(landmarks.data, out.dir, downsample.to = 1000)
}
\arguments{
\item{landmarks.data}{The landmarks data, as returned by \code{\link{load_landmarks}} or \code{\link{load_landmarks_from_dir}}}

\item{out.dir}{The output directory}

\item{downsample.to}{The number of events to be used as downsampling target}
}
\description{
Write landmarks data
}
