% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{get_highest_scoring_edges}
\alias{get_highest_scoring_edges}
\title{Add information about edge types and highest scoring edges}
\usage{
get_highest_scoring_edges(G)
}
\arguments{
\item{G}{The input \code{igraph} object}
}
\value{
Returns \code{G} with an added edge property (\code{edge_type}), and an added vertex property
  (\code{highest_scoring_edge}), with the following meaning:
  \itemize{
    \item{\code{edge_type}}: this is a string with three possible values: \code{"cluster_to_landmark"},
      \code{"inter_cluster"}, \code{"highest_scoring"}. These indicate where the edge is between a cluster and a landmark,
      between two clusters, or the highest scoring edge among all the \code{"cluster_to_landmark"}
      edges of a single cluster, respectively
    \item{\code{highest_scoring_edge}}: a number representing the index of the \code{"highest_scoring"} edge for this vertex
  }
}
\description{
Add information about edge types and highest scoring edges
}
