% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph.R
\name{get_unsupervised_graph_from_files}
\alias{get_unsupervised_graph_from_files}
\title{Builds an unsupervised graph starting from a list of input files}
\usage{
get_unsupervised_graph_from_files(files.list, metadata.tab = NULL,
  metadata.filename.col = NULL, use.basename = T, ...)
}
\arguments{
\item{files.list}{The list of files to process. The function will first determine the set of columns that are common to all
the files. Only the common vertex properties will feature in the output graph}

\item{metadata.tab}{Optional. If specified, a table of file-level metadata, to be added as vertex properties in the graph. Each row
should specify metadata for a single file, with the columns of \code{metadata.tab} representing metadata values. All the vertices
derived from that file will have the corresponding metadatata value}

\item{metadata.filename.col}{The name of the column in \code{metadata.tab} that contains the file name to be matched to the files
in \code{files.list}}

\item{use.basename}{The resulting graph will contain an additional vertex property called \code{file_name} identifying which file
the vertex was derived from. If \code{use.basename} is \code{TRUE} the \code{basename} will be used, otherwise the full path
as specified in \code{files.list}. Moreover if \code{use.basename} is \code{TRUE} the matching with the file names contained
in \code{metadata.tab} will be based on the \code{basename} only}

\item{...}{Arguments passed on to \code{get_unsupervised_graph}
\describe{
  \item{col.names}{A character vector indicating which columns of \code{tab} should be used to calculate distances}
  \item{filtering.threshold}{The threshold used to filter edges in the graph}
}}
}
\description{
This function is similar to \code{get_unsupervised_graph}, except that the input in this case is a list of file names containing
the vertices of the graph. Each file should be a tab-separated table of nodes, similar to the \code{tab} input of
\code{get_unsupervised_graph}. This function will \code{rbind} all the files in \code{files.list}, and a build a single graph
with the union of all the nodes. Optionally, file-level metadata can also be incorporated in the vertex properties of the resulting
graph, using the \code{metadata.tab} parameter
}
