% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{write_clusters_data}
\alias{write_clusters_data}
\title{Write clustering output}
\usage{
write_clusters_data(clustered.data, base.name, output.dir = "./")
}
\arguments{
\item{clustered.data}{A \code{data.frame} containing the clustered data. Each row corresponds to cell. The \code{data.frame}
must include a column called \code{cellType}, indicating cluster membership}

\item{base.name}{The base name for naming output files. This is only used if \code{clustered_data} does not contain
a column called \code{sample}. Otherwise the \code{sample} names are used for this purpose}

\item{output.dir}{The output directory}
}
\description{
Write clustering output
}
